requirements.txt
absl-py==0.13.0
certifi==2021.5.30
chardet==4.0.0
humanize==3.10.0
idna==2.10
lxml==4.6.3
mock==4.0.3
multitasking==0.0.9
numpy==1.21.0
pandas==1.3.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
scipy==1.7.0
six==1.15.0
urllib3==1.26.6
